webpage/requirements.txt
Flask==2.0.1
gunicorn==20.1.0
mysql-connector-python==8.0.26